Big commit
- many to many table for courses and students,
- UUID (GUID) instead of auto incremental IDs
- confirmation on deletes
- login errors fixed
- new error responses in grades and courses
- maybe something else I forgot

+++ routes/grades_controller.py
@@ -3,6 +3,7 @@ from shared import db
 from .models.Grade import Grade
 from .models.Student import Student
 from .models.Teacher import Teacher
+from .models.Course import Course
 
 gradesroute_blueprint = Blueprint('gradesroute_blueprint', __name__)
 
@@ -16,19 +17,22 @@ def get_grades():
         student = Student.query.get(grade.id_student)
         student_data = {"id": student.id,
                 "name": student.name,
-                "surname": student.surname,
-                "course": student.course}
+                "surname": student.surname}
     
         teacher = Teacher.query.get(grade.id_teacher)
         teacher_data = {"id": teacher.id, 
                     "name": teacher.name, 
-                    "surname": teacher.surname, 
-                    "course": teacher.course}
+                    "surname": teacher.surname}
+        
+        course = Course.query.get(grade.course)
+        
+        course_data = {"id": course.id,
+                       "name": course.name}
         
         grade_data = {"id": grade.id,
                       "id_student": student_data,
                       "id_teacher": teacher_data,
-                      "course": grade.course,
+                      "course": course_data,
                       "grade": grade.grade}
         output.append(grade_data)
         
@@ -43,77 +47,104 @@ def get_grade(id):
     
     student = Student.query.get(grade.id_student)
     student_data = {"id": student.id,
+                    "username": student.username,
                     "name": student.name,
-                    "surname": student.surname,
-                    "course": student.course}
+                    "surname": student.surname}
+    
+    course = Course.query.get(grade.course)
+        
+    course_data = {"id": course.id,
+                   "name": course.name}
     
     teacher = Teacher.query.get(grade.id_teacher)
     teacher_data = {"id": teacher.id, 
+                    "username": teacher.username,
                     "name": teacher.name, 
-                    "surname": teacher.surname, 
-                    "course": teacher.course}
+                    "surname": teacher.surname}
     
     return {"id": grade.id,
             "id_student": student_data,
             "id_teacher": teacher_data,
-            "course": grade.course,
+            "course": course_data,
             "grade": grade.grade}
 
 
 @gradesroute_blueprint.route('/grades', methods=['POST'])
 def post_grade():
+    try:
+        student = Student.query.get(request.json['student'])
+        print(student.disabled)
+        if student is None:
+            return {"message": "Student with that ID does not exist"}, 400
+        if student.disabled is True:
+            return {"message": "Student is disabled"}, 400
+    except:
+        return {"message": "JSON requires student"}
     
-    student = Student.query.get(request.json['id_student'])
+    try:
+        teacher = Teacher.query.get(request.json['teacher'])
+        if teacher is None:
+            return {"message": "Teacher with that ID does not exist"}, 400
+    except:
+        return {"message": "JSON requires teacher"}
     
-    if student is None:
-        return {"message": "Student with that ID does not exist"}, 400
+    try:
+        course = Course.query.get(request.json['course'])
+        if course is None:
+            return {"message": "Course with that ID doe no exist"}, 400
+    except:
+        return {"message": "JSON requires course"}
+    try:
+        if request.json['grade'] > 10 or request.json['grade'] <= 0:
+            return {"message": "Grades cannot be bigger than 10 and less or equal than 0"}, 400
+    except:
+        return {"message": "JSON requires grade"}
     
-    teacher = Teacher.query.get(request.json['id_teacher'])
-    
-    if teacher is None:
-        return {"message": "Teacher with that ID does not exist"}, 400
-    
-    # TODO check is course exists
-    
-    if len(request.json['course']) > 30:
-        return {"message": "course cannot be longer than 30 characters"}, 400
-    
-    if request.json['grade'] > 10 or request.json['grade'] <= 0:
-        return {"message": "grades cannot be bigger than 10 and less or equal than 0"}, 400
-    
-    grade = Grade(id_student=request.json['id_student'], 
-                  id_teacher=request.json['id_teacher'], 
+    grade = Grade(id_student=request.json['student'], 
+                  id_teacher=request.json['teacher'],
                   course=request.json['course'],
                   grade=request.json['grade'])
     db.session.add(grade)
     db.session.commit()
-    
+    course_data = {"id": course.id,
+                   "name": course.name
+                   }
     student_data = {"id": student.id,
                     "name": student.name,
-                    "surname": student.surname,
-                    "course": student.course}
+                    "surname": student.surname}
     
     teacher_data = {"id": teacher.id, 
                     "name": teacher.name, 
-                    "surname": teacher.surname, 
-                    "course": teacher.course}
+                    "surname": teacher.surname}
     
     return {"id": grade.id, 
             "student": student_data, 
-            "teacher": teacher_data, 
-            "course": grade.course, 
+            "teacher": teacher_data,
+            "course": course_data,
             "grade": grade.grade}, 201
 
 
 @gradesroute_blueprint.route('/grades/<id>', methods=['PUT'])
 def put_grade(id):
     grade = Grade.query.get(id)
+    
     if grade is None:
         return {"message": "Grade not Found"}, 404
+    
+    try:
+        course = Course.query.get(request.json['course'])
+        if course is None:
+            return {"message": "course with that ID does not exist"}, 400
+        grade.course = request.json['course']
+    except:
+        return {"message": "JSON must contain course"}, 400
+    
     try:
         student = Student.query.get(request.json['id_student'])
         if student is None:
             return {"message": "Student with that ID does not exist"}, 400
+        if student.disabled is True:
+            return {"message": "Student is disabled"}, 400
         grade.id_student = request.json['id_student']
     except:
         return {"message": "JSON must containt id_student"}, 400
@@ -131,23 +162,42 @@ def put_grade(id):
             return {"message": "grades cannot be bigger than 10 and less or equal than 0"}, 400
         grade.grade = request.json['grade']
     except:
-        return {"message": "JSON requires to have course"}, 400
+        return {"message": "JSON requires to have grade"}, 400
     
     db.session.commit()
     
+    course_data = {"id": course.id,
+                   "name": course.name
+                   }
+    student_data = {"id": student.id,
+                    "name": student.name,
+                    "surname": student.surname}
+    
+    teacher_data = {"id": teacher.id, 
+                    "name": teacher.name, 
+                    "surname": teacher.surname}
+    
     return {"id": grade.id, 
-            "id_student": grade.id_student, 
-            "id_teacher": grade.id_teacher, 
-            "course": grade.course, 
+            "student": student_data, 
+            "teacher": teacher_data, 
+            "course": course_data, 
             "grade": grade.grade}
 
 
 @gradesroute_blueprint.route('/grades/<id>', methods={'DELETE'})
 def delete_grade(id):
-	grade = Grade.query.get(id)
-	if grade is None:
-		return {"message": "grade not Found"}, 404
-	db.session.delete(grade)
-	db.session.commit()
-	return {"message": "grade deleted successfully"}, 204
+    try:
+        if request.json['confirm'] == True:
+            grade = Grade.query.get(id)
+            if grade is None:
+                return {"message": "Grade not found"}, 404
+            
+            db.session.delete(grade)
+            db.session.commit()
+            return {"message": "Grade successfully deleted"}, 200
+        else:
+            return {"message": "Confirmation needed"}, 400
+    except:
+        return {"message": "Confirmation needed"}, 400
+
